+++ apps/web/src/constants/ui-labels.ts
@@ -1,0 +1,10 @@
+/**
+ * Shared UI labels and icons to maintain consistency across the app
+ */
+
+export const UI_LABELS = {
+  EVAL_EDITOR: {
+    label: "Eval Editor",
+    description: "Edit and manage evaluations for this document",
+  },
+} as const;
